v0.3.1143 follow-on — SCALE-SEAM ⓴ ⓵ ⓶, clash import / jobs / projects
Eleven methods out of client.ts (1,740 → 1,663). No extra withX().
Also drop the unused LiveStream import that clean mobile tsc rejected
after notificationStream left client.ts.
